# Aryaa_Bhadane/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Library Management REST API
  description: Full OpenAPI 3.0 specification for Library Management System REST API built with Node.js, Express, JWT, bcrypt, and Firebase Firestore.
  version: 1.0.0
  contact:
    name: Aryaa Bhadane
servers:
  - url: /
    description: Current Server
paths:
...
  /api/books:
    get:
      summary: List all books with optional filters
      tags:
        - Books
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter by book category
        - in: query
          name: status
          schema:
            type: string
            enum: [available, borrowed]
          description: Filter by book availability status
        - in: query
          name: author
          schema:
            type: string
          description: Filter by author name
      responses:
        '200':
          description: List of books matching criteria
    post:
      summary: Add a new book (Librarian only)
      tags:
        - Books
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: Book added
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Librarian role required)
...
  /api/transactions:
    get:
      summary: View all library transactions (Librarian only)
      tags:
        - Transactions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of all library transactions
        '403':
          description: Forbidden
...
  /api/users:
    get:
      summary: List all users (Librarian only)
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of registered users (passwords omitted)
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token
  schemas:
    BookInput:
      type: object
      required:
        - title
        - author
        - isbn
        - category
        - quantity
      properties:
        title:
          type: string
          example: The Great Gatsby
        author:
          type: string
          example: F. Scott Fitzgerald
        isbn:
          type: string
          example: 978-0743273565
        category:
          type: string
          example: Fiction
        quantity:
          type: integer
          minimum: 0
          example: 5
        status:
          type: string
          enum: [available, borrowed]
          example: available